Remove SharedRegion algorithm
This worked in the prototype but the system has become
much more complex since then.

Doing this in a polymorphic fashion was quite ugly.

It had some fiddly bug.

The prototype performance was not particularly good.

We can actually achieve parallelism quite trivially while
rendering movies: by rendering frames concurrently.  This
would be easier to achieve with a non-threaded single frame
render.